refactor: shared ProviderSnapshot pipeline for popover and dashboard
One snapshot/limit model and one builder replace the two near-identical
private type pairs in MenuBarView and UsageDashboardView. Fixes drift the
duplication had accumulated: the third-limit label rule was inverted
between the two copies (canonical rule: Claude Code -> Sonnet, everything
else -> Code Review), the dashboard hero derived its icon by
string-matching its own rendered title, and pace context was recovered by
string-matching display titles instead of the limit's kind.

Genuinely shared components (UsageBar, ProviderLogoView/Kind/cache,
ExtraUsageStatusPill, the reset-countdown family) move out of
MenuBarView.swift - which had accidentally become the design system - into
Views/Components/. Builder behavior is covered by ProviderSnapshotTests.
