+++ MeterBar/Views/UsageDashboardView.swift
@@ -71,6 +71,9 @@ struct UsageDashboardView: View {
     @StateObject private var costTracker = CostTracker.shared
     @StateObject private var claudeAccountStore = ClaudeCodeAccountStore.shared
     @StateObject private var providerVisibility = ProviderVisibilityStore.shared
+    @StateObject private var claudeCodeService = ClaudeCodeLocalService.shared
+    @StateObject private var codexCliService = CodexCliLocalService.shared
+    @StateObject private var cursorService = CursorLocalService.shared
 
     @State private var selectedSection: DashboardSection = .overview
     @State private var columnVisibility = NavigationSplitViewVisibility.all
@@ -173,7 +176,8 @@ struct UsageDashboardView: View {
             DashboardStatusHero(
                 title: overviewStatusTitle,
                 detail: overviewStatusDetail,
-                color: tightestWindowColor
+                iconName: overviewStatusIconName,
+                color: overviewBand?.color ?? .secondary
             )
 
             LazyVGrid(columns: overviewGridColumns, alignment: .leading, spacing: 12) {
@@ -218,11 +222,7 @@ struct UsageDashboardView: View {
                 }
             } else {
                 ForEach(providerSnapshots) { snapshot in
-                    ProviderLimitsCard(
-                        snapshot: snapshot,
-                        accentColor: color(for: snapshot.service),
-                        updatedText: "Updated \(relativeDate(snapshot.lastUpdated))"
-                    )
+                    ProviderLimitsCard(snapshot: snapshot)
                 }
             }
         }
@@ -307,35 +307,19 @@ struct UsageDashboardView: View {
             .frame(maxWidth: .infinity, minHeight: 520, alignment: .leading)
     }
 
-    private var providerSnapshots: [DashboardProviderSnapshot] {
-        var snapshots: [DashboardProviderSnapshot] = []
-
-        if providerVisibility.isEnabled(.codexCli), let codex = dataManager.metrics[.codexCli] {
-            snapshots.append(DashboardProviderSnapshot(title: "Codex", service: .codexCli, metrics: codex))
-        }
-
-        if providerVisibility.isEnabled(.claudeCode) {
-            let claudeAccountMetrics = dataManager.claudeCodeAccountMetrics
-            if !claudeAccountMetrics.isEmpty {
-                for account in claudeAccountStore.accounts {
-                    if let metrics = claudeAccountMetrics[account.id] {
-                        snapshots.append(DashboardProviderSnapshot(
-                            title: account.isDefault && claudeAccountStore.accounts.count == 1 ? "Claude" : account.name,
-                            service: .claudeCode,
-                            metrics: metrics
-                        ))
-                    }
-                }
-            } else if let claude = dataManager.metrics[.claudeCode] {
-                snapshots.append(DashboardProviderSnapshot(title: "Claude", service: .claudeCode, metrics: claude))
-            }
-        }
-
-        if providerVisibility.isEnabled(.cursor), let cursor = dataManager.metrics[.cursor] {
-            snapshots.append(DashboardProviderSnapshot(title: "Cursor", service: .cursor, metrics: cursor))
-        }
-
-        return snapshots
+    private var providerSnapshots: [ProviderSnapshot] {
+        // Same builder the popover uses; the dashboard only renders providers
+        // that have reported metrics.
+        ProviderSnapshotBuilder.snapshots(ProviderSnapshotBuilder.Input(
+            metrics: dataManager.metrics,
+            claudeAccounts: claudeAccountStore.accounts,
+            claudeAccountMetrics: dataManager.claudeCodeAccountMetrics,
+            enabledServices: providerVisibility.enabledServices,
+            claudeCodeHasAccess: claudeCodeService.hasAccess,
+            codexCliHasAccess: codexCliService.hasAccess,
+            cursorHasAccess: cursorService.hasAccess
+        ))
+        .filter(\.hasMetrics)
     }
 
     private var visibleCostSummary: CostSummary? {
@@ -356,35 +340,37 @@ struct UsageDashboardView: View {
         return labels
     }
 
-    private var tightestWindowColor: Color {
-        guard let limit = providerSnapshots.flatMap(\.limits).min(by: { $0.percentLeft < $1.percentLeft }) else {
-            return .secondary
-        }
-        return MeterBarTheme.quotaStatusColor(percentLeft: limit.percentLeft)
+    private var tightestLimit: SnapshotLimit? {
+        providerSnapshots.tightestLimit
+    }
+
+    private var overviewBand: QuotaBand? {
+        tightestLimit.map { QuotaBand.forPercentLeft($0.percentLeft) }
     }
 
     private var overviewStatusTitle: String {
         guard !providerSnapshots.isEmpty else { return "No sources enabled" }
-        guard let limit = providerSnapshots.flatMap(\.limits).min(by: { $0.percentLeft < $1.percentLeft }) else {
-            return "Waiting for usage"
-        }
-        if limit.percentLeft <= 0 { return "Quota exhausted" }
-        if limit.percentLeft <= 10 { return "Quota needs attention" }
-        if limit.percentLeft <= 25 { return "Quota is tight" }
-        return "All tracked quotas look healthy"
+        guard let overviewBand else { return "Waiting for usage" }
+        return overviewBand.overviewTitle
+    }
+
+    private var overviewStatusIconName: String {
+        // Neutral states (no providers enabled / no usage yet) should not show
+        // the healthy green shield, which falsely implies tracked quotas look good.
+        overviewBand?.iconName ?? "circle.dashed"
     }
 
     private var overviewStatusDetail: String {
         guard !providerSnapshots.isEmpty else {
             return "Enable providers in Settings to show quota status."
         }
-        guard let limit = providerSnapshots.flatMap(\.limits).min(by: { $0.percentLeft < $1.percentLeft }) else {
+        guard let tightestLimit else {
             return "Refresh to load enabled provider status."
         }
-        if limit.percentLeft <= 0 {
-            return "\(limit.title) is out until reset. Tracking \(providerSnapshots.count) local provider sources."
+        if tightestLimit.percentLeft <= 0 {
+            return "\(tightestLimit.title) is out until reset. Tracking \(providerSnapshots.count) local provider sources."
         }
-        return "\(limit.title) has \(limit.percentLeft)% left. Tracking \(providerSnapshots.count) local provider sources."
+        return "\(tightestLimit.title) has \(tightestLimit.percentLeft)% left. Tracking \(providerSnapshots.count) local provider sources."
     }
 
     private var sectionSubtitle: String {
@@ -435,57 +421,6 @@ struct UsageDashboardView: View {
         guard activeSection == .overview || activeSection == .costs else { return }
         await costTracker.refreshMissingDaysInBackground(days: 30)
     }
-
-    private func color(for service: ServiceType) -> Color {
-        MeterBarTheme.accent(for: service)
-    }
-
-    private func relativeDate(_ date: Date) -> String {
-        UsageFormat.relative(date)
-    }
-}
-
-private struct DashboardProviderSnapshot: Identifiable {
-    let id: String
-    let title: String
-    let service: ServiceType
-    let logoKind: ProviderLogoKind
-    let lastUpdated: Date
-    let limits: [DashboardLimit]
-
-    init(title: String, service: ServiceType, metrics: UsageMetrics) {
-        self.id = "\(service.rawValue)-\(title)"
-        self.title = title
-        self.service = service
-        self.logoKind = providerLogoKind(for: service)
-        self.lastUpdated = metrics.lastUpdated
-        self.limits = [
-            DashboardLimit(title: "Session", limit: metrics.sessionLimit),
-            DashboardLimit(title: "Weekly", limit: metrics.weeklyLimit),
-            DashboardLimit(title: service == .codexCli ? "Code Review" : "Sonnet", limit: metrics.codeReviewLimit)
-        ].compactMap { $0 }
-    }
-}
-
-private struct DashboardLimit: Identifiable {
-    let id = UUID()
-    let title: String
-    let usageLimit: UsageLimit
-
-    init?(title: String, limit: UsageLimit?) {
-        guard let limit else { return nil }
-        self.title = title
-        self.usageLimit = limit
-    }
-
-    var usedPercent: Double {
-        usageLimit.rawPercentage
-    }
-
-    var percentLeft: Int {
-        let remainingPercent = max(0, 100 - usedPercent)
-        return remainingPercent == 0 ? 0 : max(1, Int(ceil(remainingPercent)))
-    }
 }
 
 private let overviewTileMinHeight: CGFloat = 220
@@ -493,24 +428,8 @@ private let overviewTileMinHeight: CGFloat = 220
 private struct DashboardStatusHero: View {
     let title: String
     let detail: String
+    let iconName: String
     let color: Color
-
-    private var iconName: String {
-        if title.localizedCaseInsensitiveContains("exhausted") {
-            return "exclamationmark.octagon.fill"
-        }
-        if title.localizedCaseInsensitiveContains("attention")
-            || title.localizedCaseInsensitiveContains("tight") {
-            return "exclamationmark.triangle.fill"
-        }
-        // Neutral states (no providers enabled / no usage yet) should not show the
-        // healthy green shield, which falsely implies tracked quotas look good.
-        if title.localizedCaseInsensitiveContains("no sources")
-            || title.localizedCaseInsensitiveContains("waiting") {
-            return "circle.dashed"
-        }
-        return "checkmark.shield.fill"
-    }
 
     var body: some View {
         HStack(alignment: .center, spacing: 14) {
@@ -540,33 +459,25 @@ private struct DashboardStatusHero: View {
 }
 
 private struct ProviderOverviewStatusCard: View {
-    let snapshot: DashboardProviderSnapshot
-
-    private var accentColor: Color {
-        MeterBarTheme.accent(for: snapshot.service)
-    }
-
-    private var primaryLimit: DashboardLimit? {
-        snapshot.limits.min { $0.percentLeft < $1.percentLeft }
-    }
+    let snapshot: ProviderSnapshot
 
     private var statusText: String {
-        guard let primaryLimit else { return "No data" }
-        if primaryLimit.percentLeft <= 0 { return "Out" }
-        if primaryLimit.percentLeft <= 10 { return "Critical" }
-        if primaryLimit.percentLeft <= 25 { return "Tight" }
-        return "Healthy"
+        snapshot.band?.shortLabel ?? "No data"
+    }
+
+    private var statusColor: Color {
+        snapshot.band?.color ?? .secondary
     }
 
     var body: some View {
         VStack(alignment: .leading, spacing: 12) {
             HStack(alignment: .center, spacing: 9) {
-                ProviderLogoView(kind: snapshot.logoKind, size: 20, foregroundColor: accentColor)
+                ProviderLogoView(kind: snapshot.logoKind, size: 20, foregroundColor: snapshot.accentColor)
                 VStack(alignment: .leading, spacing: 2) {
                     Text(snapshot.title)
                         .font(.headline)
                         .fontWeight(.semibold)
-                    Text("Updated \(relativeDate(snapshot.lastUpdated))")
+                    Text(updatedText)
                         .font(.caption)
                         .foregroundColor(.secondary)
                 }
@@ -585,7 +496,7 @@ private struct ProviderOverviewStatusCard: View {
             } else {
                 VStack(alignment: .leading, spacing: 12) {
                     ForEach(snapshot.limits) { limit in
-                        DashboardLimitRow(limit: limit, accentColor: accentColor)
+                        DashboardLimitRow(limit: limit, accentColor: snapshot.accentColor)
                     }
                 }
             }
@@ -599,13 +510,9 @@ private struct ProviderOverviewStatusCard: View {
         .dashboardCardBackground()
     }
 
-    private var statusColor: Color {
-        guard let primaryLimit else { return .secondary }
-        return MeterBarTheme.quotaStatusColor(percentLeft: primaryLimit.percentLeft)
-    }
-
-    private func relativeDate(_ date: Date) -> String {
-        UsageFormat.relative(date)
+    private var updatedText: String {
+        guard let updatedAt = snapshot.updatedAt else { return "No data" }
+        return "Updated \(UsageFormat.relative(updatedAt))"
     }
 }
 
@@ -694,9 +601,7 @@ private struct CostOverviewStatusCard: View {
 }
 
 private struct ProviderLimitsCard: View {
-    let snapshot: DashboardProviderSnapshot
-    let accentColor: Color
-    let updatedText: String
+    let snapshot: ProviderSnapshot
 
     var body: some View {
         VStack(alignment: .leading, spacing: 12) {
@@ -704,7 +609,7 @@ private struct ProviderLimitsCard: View {
                 ProviderTitle(
                     title: snapshot.title,
                     logoKind: snapshot.logoKind,
-                    color: accentColor,
+                    color: snapshot.accentColor,
                     font: .title3
                 )
                 Spacer()
@@ -719,13 +624,18 @@ private struct ProviderLimitsCard: View {
                     .foregroundColor(.secondary)
             } else {
                 ForEach(snapshot.limits) { limit in
-                    DashboardLimitRow(limit: limit, accentColor: accentColor)
+                    DashboardLimitRow(limit: limit, accentColor: snapshot.accentColor)
                 }
             }
         }
         .padding(14)
         .frame(maxWidth: .infinity, alignment: .topLeading)
         .dashboardCardBackground()
+    }
+
+    private var updatedText: String {
+        guard let updatedAt = snapshot.updatedAt else { return "No data" }
+        return "Updated \(UsageFormat.relative(updatedAt))"
     }
 }
 
@@ -778,12 +688,8 @@ private struct ProviderTitle: View {
 }
 
 private struct DashboardLimitRow: View {
-    let limit: DashboardLimit
+    let limit: SnapshotLimit
     let accentColor: Color
-
-    private var paceContext: PaceLabelContext {
-        limit.title.localizedCaseInsensitiveContains("weekly") ? .weekly : .session
-    }
 
     private var isOut: Bool {
         limit.percentLeft <= 0
@@ -806,7 +712,7 @@ private struct DashboardLimitRow: View {
                 usedPercentage: limit.usedPercent,
                 accentColor: accentColor,
                 pace: limit.usageLimit.pace(),
-                paceContext: paceContext
+                paceContext: limit.paceContext
             )
 
             HStack {
@@ -1419,9 +1325,9 @@ private struct DailyProviderUsageSummaryRow: View {
     var body: some View {
         HStack(spacing: 10) {
             ProviderLogoView(
-                kind: providerLogoKind(for: provider.provider),
+                kind: .forService(provider.provider),
                 size: 14,
-                foregroundColor: color(for: provider.provider)
+                foregroundColor: MeterBarTheme.accent(for: provider.provider)
             )
             Text(provider.provider.displayName)
                 .font(.caption)
@@ -1442,7 +1348,7 @@ private struct ProviderCostBreakdown: View {
     let cost: TokenCost
 
     private var logoKind: ProviderLogoKind {
-        providerLogoKind(for: cost.provider)
+        .forService(cost.provider)
     }
 
     private var logoColor: Color {
@@ -1583,21 +1489,6 @@ private enum DashboardDateFormat {
     static func medium(_ date: Date) -> String { mediumDate.string(from: date) }
     static func month(_ date: Date) -> String { month.string(from: date) }
     static func weekdayMonthDay(_ date: Date) -> String { weekdayMonthDay.string(from: date) }
-}
-
-private func providerLogoKind(for provider: ServiceType) -> ProviderLogoKind {
-    switch provider {
-    case .codexCli:
-        return .codex
-    case .claudeCode:
-        return .claude
-    case .cursor:
-        return .cursor
-    }
-}
-
-private func color(for provider: ServiceType) -> Color {
-    MeterBarTheme.accent(for: provider)
 }
 
 private extension View {
